Refactor ValidatorSource classes removing repetition
